// src/services/customers/types.ts
export type User = {
    firstName: string
    lastName: string
    email: string
    gender: string
    country: string
    city: string
    state: string
    postCode: string
    street: string
    streetNumber: string
}

export type UserListFilterType = {
    gender: string
    country: string
    lastName: string
}